Task 3 muokattu

+++ src/Moduuli1/Task3.java
@@ -16,11 +16,11 @@ public class Task3 {
         System.out.println("Give the third number:");
         int third = Integer.parseInt(scanner.nextLine());
 
-        float avg = (float)(first + second + third) / 3;
+        double avg = (double)(first + second + third) / 3;
 
         System.out.println("The sum of the numbers is " + (first + second + third));
         System.out.println("The product of the numbers is " +(first * second * third));
         System.out.println("The average of the numbers is " + avg);
-        //Int type tulostaa ainoastaan kokonaislukuja. Float tai double tulostaa desimaalien kanssa
+        //Int type tulostaa ainoastaan kokonaislukuja. double tulostaa desimaalit. Myös Float.
     }
 }
